refactor: split src/http.rs into route surface modules
Split the monolithic src/http.rs (4517 lines) into focused submodules
under src/http/ to improve navigability and maintainability.

Module breakdown:
- mod.rs: AppState, router(), serve_unix(), DTOs, shared helpers, tests
- cors.rs: CORS layer configuration
- error.rs: HttpErrorEnvelope and error response helpers
- runtime.rs: runtime status/readiness/config/credentials handlers
- agents.rs: agent listing, status, briefs, transcript, skills handlers
- state.rs: agent state snapshot and bootstrap helpers
- tasks.rs: task lifecycle handlers (status/output/input/stop/create)
- work_items.rs: work item CRUD handlers
- timers.rs: timer listing and lifecycle handlers
- events.rs: events, event stream, and message handlers
- control.rs: control, abort, workspace, model, prompt, transport handlers
- enqueue.rs: message enqueue handlers

No API behavior changes. All routes, request/response shapes, auth, and
CORS behavior are preserved. Handler functions are re-exported from
mod.rs via pub(crate) use to keep the router composition unchanged.

Update http_route_snapshot test to read src/http/mod.rs instead of
the old single-file path.

Closes #1896.
